+++ sources/S8-53/Device53/Display/DisplayPrimitives.cpp
@@ -13,7 +13,7 @@ void ProgressBar_Draw(ProgressBar *bar)
     sprintf(buffer, "Завершено %.1f %%", passedPercents);
     painter.DrawStringInCenterRectC(x, y - 15, bar->width, bar->height, buffer, COLOR_FILL);
     painter.DrawRectangle(bar->x, bar->y, bar->width, bar->height);
-    painter.FillRegion(bar->x, bar->y, bar->width * passedPercents / 100.0f, bar->height);
+    painter.FillRegion(bar->x, bar->y, (int)(bar->width * passedPercents / 100.0f), bar->height);
     buffer[0] = 0;
     sprintf(buffer, "Осталось %.1f с", (int)(bar->fullTime - bar->passedTime) / 1000.0f);
     painter.DrawStringInCenterRect(x, y + bar->height, bar->width, bar->height, buffer);
